# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/account/preferences": {
      "get": {
        "operationId": "AccountController_preferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferencesDto"
                }
              }
            }
          }
        },
        "tags": ["Account"]
      },
      "post": {
        "operationId": "AccountController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferencesDto"
                }
              }
            }
          }
        },
        "tags": ["Account"]
      }
    },
...
  },
  "info": {
    "title": "Cat Care API",
    "description": "Cat Care service contract.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AccountPreferencesDto": {
        "type": "object",
        "properties": {
          "themePreference": {
            "type": "string",
            "enum": ["system", "light", "dark"]
          },
          "languagePreference": {
            "type": "string",
            "enum": ["system", "en", "pl"]
          }
        },
        "required": ["themePreference", "languagePreference"]
      },
      "UpdateAccountPreferencesDto": {
        "type": "object",
        "properties": {
          "themePreference": {
            "type": "string",
            "enum": ["system", "light", "dark"]
          },
          "languagePreference": {
            "type": "string",
            "enum": ["system", "en", "pl"]
          }
        }
      },
...
    }
  }
}